// src/common/config/types/enviroment.enum.ts
export enum EnvEnum {
  NODE_ENV = 'NODE_ENV',
  PORT = 'PORT',
  LOG_LEVEL = 'LOG_LEVEL',
  CLIENT_TOKEN = 'CLIENT_TOKEN',
  DISCORD_APP_ID = 'DISCORD_APP_ID',
  DB_TYPE = 'DB_TYPE',
  DB_PORT = 'DB_PORT',
  DB_USER = 'DB_USER',
  DB_PASS = 'DB_PASS',
  DB_NAME = 'DB_NAME',
  DB_HOST = 'DB_HOST',
  DB_SCHEMA = 'DB_SCHEMA',
  DB_SYNCHRONIZE = 'DB_SYNCHRONIZE',
  DB_LOGGING = 'DB_LOGGING',
  DB_LOGGER_LEVEL = 'DB_LOGGER_LEVEL',
  DB_PRISMA_CONNECTION = 'DB_PRISMA_CONNECTION',
}